+++ capstone_flutter/lib/view/screen/wifi_screen/ilustration_success_wifi_screen.dart
@@ -9,8 +9,8 @@ class IlustrationSuccessWifi extends StatefulWidget {
   final String pelangganData;
   final DateTime createdAt;
   final String providerName;
-  final int price;
-  final int adminFee;
+  final double price;
+  final double adminFee;
   final String customerName;
   const IlustrationSuccessWifi(
       {super.key,
